Enhance the logging mechanism a bit
For Copr usecase, it is easier to not pass "log_level" down to
"get_rpms_to_remove" method, but the logger directly (the one we have in
copr is the one that is sending log records to Redis).

Also, not all the methods needed the `dry_run` option.

+++ prunerepo/helpers.py
@@ -10,14 +10,12 @@ import logging
 
 from prunerepo.pair_srpm_rpm import RPMToSRPMPairs
 
-log = logging.getLogger(__name__)
-
 
 def is_srpm(package):
     return package.endswith(".src.rpm")
 
 
-def rm_file(path, dry_run):
+def rm_file(path, dry_run, log):
     """
     Remove file given its absolute path
     """
@@ -28,13 +26,13 @@ def rm_file(path, dry_run):
         os.remove(path)
 
 
-def run_cmd(cmd, dry_run):
+def run_cmd(cmd, log, dry_run=False):
     """
     Run given command in a subprocess
     """
     log.debug("Executing: " + ' '.join(cmd))
     if dry_run:
-        return
+        return []
     process = subprocess.Popen(cmd, stdout=subprocess.PIPE, stderr=subprocess.PIPE)
     (stdout, stderr) = process.communicate()
     sys.stderr.write(stderr.decode(encoding='utf-8'))
@@ -43,54 +41,52 @@ def run_cmd(cmd, dry_run):
     return stdout.decode(encoding='utf-8').splitlines()
 
 
-def get_package_build_time(package_path, dry_run):
+def get_package_build_time(package_path, log):
     """
     Get build time by reading package metadata
     """
     query_cmd = ["/usr/bin/rpm", "-qp", "--queryformat", "%{BUILDTIME}"] + [package_path]
-    stdout = run_cmd(query_cmd, dry_run)
+    stdout = run_cmd(query_cmd, log)
     return int(stdout[0])
 
 
-def get_rpms(repoquery_cmd, path, dry_run):
+def get_rpms(repoquery_cmd, path, log):
     """
     Get paths to rpm packages in the repository according to given repoquery_cmd
     """
-    stdout = run_cmd(repoquery_cmd, dry_run)  # returns srpms as well
+    stdout = run_cmd(repoquery_cmd, log)  # returns srpms as well
     rel_rpms_paths = [relpath for relpath in stdout if not is_srpm(relpath)]
     abs_rpms_paths = [os.path.abspath(os.path.join(path, relpath)) for relpath in rel_rpms_paths]
     return abs_rpms_paths
 
 
-def prune_packages(path, days, log_level, dry_run):
+def prune_packages(path, days, dry_run, log):
     """
     Remove obsoleted packages
     """
-    if not set_logging_level(log_level):
-        sys.exit(1)
     log.debug('Removing obsoleted packages...')
-    rpms = get_rpms_to_remove(path, log_level, days)
+    rpms = get_rpms_to_remove(path, days, log)
     was_deletion = False
     if not rpms:
         log.error("No RPMs available")
         return was_deletion
     for rpm in rpms:
-        rm_file(rpm, dry_run)
+        rm_file(rpm, dry_run, log)
         was_deletion = True
     return was_deletion
 
 
-def recreate_repo(path, dry_run):
+def recreate_repo(path, dry_run, log):
     """
     Recreate the repository by using createrepo_c
     """
     log.debug("Recreating repository...")
     createrepo_cmd = ['/usr/bin/createrepo_c', '--database', '--update', '--local-sqlite',
                       '--cachedir', '/tmp/', '--workers', '8'] + [path]
-    return run_cmd(createrepo_cmd, dry_run)
+    return run_cmd(createrepo_cmd, log, dry_run)
 
 
-def clean_copr(path, days, dry_run):
+def clean_copr(path, days, dry_run, log):
     """
     Remove whole copr build dirs if they no longer contain a srpm/rpm file
     """
@@ -110,41 +106,41 @@ def clean_copr(path, days, dry_run):
             continue
 
         log.info('Removing: ' + dir_path)
-        shutil.rmtree(dir_path)
+        if not dry_run:
+            shutil.rmtree(dir_path)
 
         # also remove the associated log in the main dir
         build_id = os.path.basename(dir_path).split('-')[0]
         buildlog_name = 'build-' + build_id + '.log'
         buildlog_path = os.path.abspath(os.path.join(path, buildlog_name))
-        rm_file(os.path.join(path, buildlog_path), dry_run)
+        rm_file(os.path.join(path, buildlog_path), dry_run, log)
 
 
-def set_logging_level(log_level):
+def get_logger(log_level="INFO", module=None):
     """
     Set logging level
     """
+    log = logging.getLogger(module or __name__)
+    if log.handlers:
+        # repeated call
+        return log
+
     handler = logging.StreamHandler(stream=sys.stderr)
     handler.setFormatter(logging.Formatter("%(message)s"))
-    try:
-        handler.setLevel(log_level.upper())
-        log.addHandler(handler)
-        log.setLevel(handler.level)
-        return True
-    except ValueError as error:
-        print(str(error), file=sys.stderr)
-    except TypeError as error:
-        print(str(error), file=sys.stderr)
-    return False
+    handler.setLevel(log_level.upper())
+    log.addHandler(handler)
+    log.setLevel(handler.level)
+    return log
 
 
-def get_rpms_to_remove(directory, log_level='INFO', days=0):
+def get_rpms_to_remove(directory, days=0, log=None):
     """
     Returns a list of (s)rpm path names that should be removed.
     0 days means that (s)rpm will be removed regardless of when the package was built.
 
     :param directory: local path to a yum repository
-    :param log_level: set logging to desired level (error, info or debug)
     :param days: how old are the packages to be removed, in the number of days
+    :param log: logger to use, if not specified an INFO stderr logger is created
     :return: a list of (s)RPM path names that should be removed
     """
     get_all_packages_cmd = [
@@ -157,15 +153,19 @@ def get_rpms_to_remove(directory, log_level='INFO', days=0):
         "--quiet",
         "--setopt=skip_if_unavailable=False",
     ]
-    if not log.handlers:
-        if not set_logging_level(log_level):
-            return []
-    log.info("Checking '%s' directory for removal candidates older than %s days" % (os.path.abspath(directory), days))
+
+    if not log:
+        log = get_logger()
+
+    log.info("Checking '%s' repo for removal candidates older than %s days",
+             os.path.abspath(directory), days)
+
     get_latest_packages_cmd = get_all_packages_cmd + ['--latest-limit=1']
-    latest_rpms = get_rpms(get_latest_packages_cmd, directory, dry_run=False)
+    latest_rpms = get_rpms(get_latest_packages_cmd, directory, log)
     if not latest_rpms:
         return []
-    all_rpms = get_rpms(get_all_packages_cmd, directory, dry_run=False)
+
+    all_rpms = get_rpms(get_all_packages_cmd, directory, log)
 
     repodir = os.path.abspath(directory)
     pair_lookup = RPMToSRPMPairs(repodir, log)
@@ -173,8 +173,8 @@ def get_rpms_to_remove(directory, log_level='INFO', days=0):
     to_remove_rpms = set(all_rpms) - set(latest_rpms)
     rpm_list = []
     for rpm in to_remove_rpms:
-        log.debug("Checking age of the '%s' file" % os.path.split(rpm)[1])
-        if time.time() - get_package_build_time(rpm, dry_run=False) < days * 24 * 3600:
+        log.debug("Checking age of the '%s' file", os.path.split(rpm)[1])
+        if time.time() - get_package_build_time(rpm, log) < days * 24 * 3600:
             continue
         rpm_list.append(rpm)
